# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Constituent Touchpoint Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
    post:
      operationId: createUser
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /users/{id}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordResetInput" }
      responses:
        "204": { description: Done }

  # ───────── Campaign types
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, active, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [standard, admin, super_admin] }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
    UserInput:
      type: object
      required: [email, name, role, password]
      properties:
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [standard, admin, super_admin] }
        password: { type: string, minLength: 8 }
    UserUpdate:
      type: object
      properties:
        name: { type: string }
        role: { type: string, enum: [standard, admin, super_admin] }
        active: { type: boolean }
    PasswordResetInput:
      type: object
      required: [password]
      properties:
        password: { type: string, minLength: 8 }
...